test: LottoTest 를 추가하고 WinnerLottoTest 의 테스트코드를 이전함.
